Hospital Management/Hospital Management/requirements.txt
Flask==2.1.0
Flask-PyMongo==2.3.0
Flask-MySQLdb==1.0.1
